[#28] refactor : 패키지 재 구성
(cherry picked from commit 5cb2a269d69cc135a4aac16afbe5c84b863bc4d4)
