+++ app/src/main/java/com/euphony/project/account_touch/utils/model/ExternalPackage.kt
@@ -1,11 +1,10 @@
-package com.euphony.project.account_touch.data.entity.model
+package com.euphony.project.account_touch.utils.model
 import com.euphony.project.account_touch.R
 
 
 enum class ExternalPackage(val packageId: Int){
     TOSS(R.string.toss_package),
     KAKAOPAY(R.string.kakaopay_package),
-
     KAKAOBANK(R.string.kakaobank_package),
     KOOKMIN(R.string.kookmin_package),
     KBANK(R.string.kbank_package),
